moved string_parsing up a directory
